Minor fixes to IndexReaderUtils and related tests
